MVC for article page
